Reorganize actions for `icmaa-catalog` and `icmaa-user`
