+++ src/modules/icmaa-catalog/store/category/actions.ts
@@ -2,6 +2,7 @@ import Vue from 'vue'
 import { ActionTree } from 'vuex'
 import RootState from '@vue-storefront/core/types/RootState'
 import CategoryState from '@vue-storefront/core/modules/catalog-next/store/category/CategoryState'
+import FilterVariant from '@vue-storefront/core/modules/catalog-next/types/FilterVariant'
 import { Category } from '@vue-storefront/core/modules/catalog-next/types/Category'
 import { DataResolver } from '@vue-storefront/core/data-resolver/types/DataResolver'
 import { router } from '@vue-storefront/core/app'
@@ -18,6 +19,13 @@ const actions: ActionTree<CategoryState, RootState> = {
   ...searchActions,
   async loadCategoryWithExtras ({ dispatch }, categorySearchOptions: DataResolver.CategorySearchOptions): Promise<Category> {
     return dispatch('icmaaCategoryExtras/loadCategoryWithExtras', categorySearchOptions, { root: true })
+  },
+  async setSearchFilters ({ dispatch }, filterVariants: FilterVariant[] = []) {
+    let currentQuery = {}
+    filterVariants.forEach(filterVariant => {
+      currentQuery = changeFilterQuery({ currentQuery, filterVariant })
+    })
+    await dispatch('changeRouterFilterParameters', currentQuery)
   },
   async unsetSearchFilterForAttribute ({ dispatch, getters }, attributeKey: string) {
     let currentQuery = router.currentRoute[products.routerFiltersSource]
